src/test/java/com/leetcode/dp/medium/PerfectSquares_279Test.java
package com.leetcode.dp.medium;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class PerfectSquares_279Test {

  private final PerfectSquares_279 solution = new PerfectSquares_279();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(12, 3), arguments(13, 2), arguments(40, 2));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void numSquares(int n, int expectedResult) {
    int actualResult = solution.numSquares(n);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
